Made ExternMFD to Work

+++ Src/Orbiter/Psys.cpp
@@ -289,7 +289,7 @@ intptr_t PlanetarySystem::FindFirst (int type, _finddata_t *fdata, char *path, c
 	return fh;
 }
 
-int PlanetarySystem::FindNext (intptr_t fh, _finddata_t *fdata, char *fname)
+intptr_t PlanetarySystem::FindNext (intptr_t fh, _finddata_t *fdata, char *fname)
 {
 	intptr_t fn = _findnext (fh, fdata);
 	if (!fn) {
